Allow for custom stream urls
